Support structured logging
A Logger can be called with the methods "WithField[s]" which will
provide structured data to the underlying handlers. It's still the
handler's responsibility about how to ultimatly format the data.

A Logger can still be used with the normal log level methods.
In which case there will be no structured data available.

This does break compatibility with v1 handelers. But logging
compatibility is not broken. The Custom log level has been removed.

TODO: Format structured data in both included handlers. Create tests
for Entry.